Cleaned up FieldOptions file
