+++ src/sdk/PnP.Core.Test/SharePoint/ListFieldTests.cs
@@ -4,6 +4,7 @@ using PnP.Core.Test.Utilities;
 using System;
 using System.Linq;
 using System.Threading.Tasks;
+using PnP.Core.Model.SharePoint.Core.Public;
 using PnP.Core.QueryModel;
 
 namespace PnP.Core.Test.SharePoint
