+++ src/sdk/PnP.Core/Model/SharePoint/Core/Public/IFieldCollection.cs
@@ -1,6 +1,7 @@
 ﻿using PnP.Core.Services;
 using System.Linq;
 using System.Threading.Tasks;
+using PnP.Core.Model.SharePoint.Core.Public;
 
 namespace PnP.Core.Model.SharePoint
 {
